Updated dependencies

+++ ledger-kafka/src/test/java/com/obsidiandynamics/blackstrom/ledger/KafkaTimeouts.java
@@ -1,6 +1,6 @@
 package com.obsidiandynamics.blackstrom.ledger;
 
 public interface KafkaTimeouts {
-  static long CLUSTER_AWAIT = 120_000;
-  static long TOPIC_CREATE = 10_000;
+  static int CLUSTER_AWAIT = 120_000;
+  static int TOPIC_CREATE = 10_000;
 }
